move http classes to right packages
