+++ src/main/java/com/mycompany/config/WebClientConfig.java
@@ -1,4 +1,4 @@
-package com.mycompany.http;
+package com.mycompany.config;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
