+++ src/main/java/com/mycompany/currency/BybitCryptocurrencyPair.java
@@ -1,6 +1,6 @@
 package com.mycompany.currency;
 
-import com.mycompany.HttpRequestFactory;
+import com.mycompany.http.HttpRequestFactory;
 import com.mycompany.json.JsonReader;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
